mpu9250|hmc5983, tipo variable de salida

+++ hmc5983/hmc5983.h
@@ -17,8 +17,8 @@ uint8_t HMC5983_GetStatus(HMC5983_HandleTypeDef *hmc5983);
 
 //*************************************************************************************//
 
-void HMC5983_GetMagnetometer(HMC5983_HandleTypeDef *hmc5983, int32_t *adcvalue);
-void HMC5983_GetTemperature(HMC5983_HandleTypeDef *hmc5983, int32_t *adcvalue);
+void HMC5983_GetMagnetometer(HMC5983_HandleTypeDef *hmc5983, int16_t *adcvalue);
+void HMC5983_GetTemperature(HMC5983_HandleTypeDef *hmc5983, int16_t *adcvalue);
 
 #ifdef __cplusplus
 }
